Update: - Created parallel routes for Room and UserList, to ease loading of data. - UserList is now updated to show the active users in the room.
