+++ src/app/chat/[id]/page.tsx
@@ -1,0 +1,4 @@
+export default async function Page(props: PageProps<'/chat/[id]'>) {
+	const { id } = await props.params
+	return <div>{id}</div>
+}
