+++ src/app/chat/layout.tsx
@@ -1,11 +1,18 @@
+import { getKindeServerSession } from '@kinde-oss/kinde-auth-nextjs/server'
+import { notFound } from 'next/navigation'
 import { UserBadgeClient } from '@/components/user/client-user-badge'
+import Container from '../components/container'
 import ContentGrid from '../components/grids/content-grid'
+import { ChatProvider } from '../contexts/chat-context'
 
-export default function RootLayout({
-	children,
-}: Readonly<{
-	children: React.ReactNode
-}>) {
+export default async function RootLayout({
+	room,
+	users,
+}: LayoutProps<'/chat'>) {
+	const user = getKindeServerSession()
+	const token = await user.getIdTokenRaw()
+	const host = process.env.PARTYKIT
+	if (!(host && token)) return notFound()
 	return (
 		<ContentGrid className="h-dvh grid-rows-[auto_1fr] pt-3 space-y-4">
 			<header className="rounded-lg p-5 bg-white border-b-2">
@@ -14,7 +21,12 @@ export default function RootLayout({
 					<UserBadgeClient />
 				</nav>
 			</header>
-			{children}
+			<ChatProvider host={host} token={token}>
+				<div className="grid grid-rows-[1fr_auto] p-0 space-y-4">
+					<Container className="flex flex-col">{room}</Container>
+					<Container className="flex gap-4">{users}</Container>
+				</div>
+			</ChatProvider>
 			<footer />
 		</ContentGrid>
 	)
